# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ScholarsToday IELTS API",
        "description": "Dokumentasi API",
        "contact": {
            "email": "kevinmajesta@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development Server"
        }
    ],
    "paths": {
        "/api/ielts/questions": {
            "get": {
                "tags": [
                    "IELTS"
                ],
                "summary": "Ambil daftar semua essay dan soal",
                "operationId": "getIeltsList",
                "responses": {
                    "200": {
                        "description": "Berhasil ambil data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/ielts/questions/{id}": {
            "get": {
                "tags": [
                    "IELTS"
                ],
                "summary": "Ambil detail soal berdasarkan ID",
                "operationId": "getIeltsDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Essay",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil ambil data"
                    },
                    "404": {
                        "description": "Essay tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/ielts/submit": {
            "post": {
                "tags": [
                    "IELTS"
                ],
                "summary": "Submit Jawaban & Hitung Skor",
                "operationId": "submitAnswers",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "answers": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "question_id": {
                                                    "type": "integer"
                                                },
                                                "option_id": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Skor dihitung"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/ielts/essays": {
            "post": {
                "tags": [
                    "IELTS Admin"
                ],
                "summary": "Tambah Soal Baru (Admin)",
                "operationId": "createEssay",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content",
                                    "questions"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "IELTS Reading Test 1"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Read the following passage and answer the questions..."
                                    },
                                    "questions": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "question_text",
                                                "options"
                                            ],
                                            "properties": {
                                                "question_text": {
                                                    "type": "string",
                                                    "example": "What is the main idea of the passage?"
                                                },
                                                "options": {
                                                    "type": "array",
                                                    "items": {
                                                        "required": [
                                                            "option_text",
                                                            "is_correct"
                                                        ],
                                                        "properties": {
                                                            "option_text": {
                                                                "type": "string",
                                                                "example": "Climate change impacts"
                                                            },
                                                            "is_correct": {
                                                                "type": "boolean",
                                                                "example": true
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "example": [
                                            {
                                                "question_text": "What is the main idea of the passage?",
                                                "options": [
                                                    {
                                                        "option_text": "Climate change impacts",
                                                        "is_correct": true
                                                    },
                                                    {
                                                        "option_text": "Global warming myths",
                                                        "is_correct": false
                                                    },
                                                    {
                                                        "option_text": "Environmental policies",
                                                        "is_correct": false
                                                    },
                                                    {
                                                        "option_text": "Renewable energy sources",
                                                        "is_correct": false
                                                    }
                                                ]
                                            },
                                            {
                                                "question_text": "According to the text, what year was mentioned?",
                                                "options": [
                                                    {
                                                        "option_text": "2020",
                                                        "is_correct": false
                                                    },
                                                    {
                                                        "option_text": "2021",
                                                        "is_correct": true
                                                    },
                                                    {
                                                        "option_text": "2022",
                                                        "is_correct": false
                                                    },
                                                    {
                                                        "option_text": "2023",
                                                        "is_correct": false
                                                    }
                                                ]
                                            }
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Soal Dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The questions field is required."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token authentication",
                "name": "Token",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}